Move digest_data into data_gatherer.
